+++ frontend/src/views/helpcenterstationview/InventorySidebarGroup.vue
@@ -50,6 +50,10 @@ const {t} = useI18n()
                    to="/helpcenter/station/inventory/edit" @navigate="close">
         {{ t('sidebar.inventoryEdit') }}
       </SidebarLink>
+      <SidebarLink :icon="['fas', 'right-left']" name="help-inventory-move"
+                   to="/helpcenter/station/inventory/move" @navigate="close">
+        {{ t('helpCenter.inventoryMove.title') }}
+      </SidebarLink>
       <SidebarLink :icon="['fas', 'tag']" name="help-inventory-item-detail"
                    to="/helpcenter/station/inventory/item/0" @navigate="close">
         {{ t('helpCenter.itemDetail.title') }}